# toteco_openapi.yaml
openapi: 3.0.2
info:
  title: TOTECO API
  description: API to manage publications about products (coffee, tea or toast) like a social media. The idea is to create a community that publishes photos of their toasts with coffee or tea adding the price and being able to rate the products and the establishments.
  version: "1.0"
  contact:
    email: a25540@svalero.com
servers:
  - url: http://localhost:8081
...
paths:
...
  /establishments:
    get:
      tags:
        - Establishment
      summary: Get all establishments
      description: Get all establishments
      operationId: getAllEstablishments
      responses:
        200:
          description: Ok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Establishment'
        401:
          description: Unauthorize exception
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
                
    post:
      security:
        - bearerAuth: []
      tags:
        - Establishment
      summary: Save establishment
      description: Save one establishment
      operationId: saveEstablishment
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EstablishmentDTO'
      responses:
        200:
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Establishment'
        401:
          description: Unauthorize exception
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
                    
    delete:
      security:
        - bearerAuth: []
      tags:
        - Establishment
      summary: Delete establishments
      description: Delete all establishments
      operationId: deleteAllEstablishments
      responses:
        200:
          description: Ok
          content:
            application/json:
              schema:
                type: string
                example: All establishments deleted.
        401:
          description: Unauthorize exception
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
  /establishments/{id}:
    get:
      tags:
        - Establishment
      summary: Get establishment by id
      description: Get one establishment by the indicated id
      operationId: getEstablishmentById
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        200:
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Establishment'
        401:
          description: Unauthorize exception
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        404:
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
    
    put:
      security:
        - bearerAuth: []
      tags:
        - Establishment
      summary: Update establishment
      description: Update an existing establishment
      operationId: updateEstablishment
      parameters:
        - $ref: '#/components/parameters/id'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EstablishmentDTO'
      responses:
        200:
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Establishment'
        401:
          description: Unauthorize exception
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        404:
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
                    
    delete:
      security:
        - bearerAuth: []
      tags:
        - Establishment
      summary: Delete establishment by id
      description: Delete an existing establishment
      operationId: deleteEstablishment
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        200:
          description: Ok
          content:
            application/json:
              schema:
                type: string
                example: Establishment deleted.
        401:
          description: Unauthorize exception
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        404:
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
...
  /establishments/date/{date}:
    get:
      tags:
        - Establishment
      summary: Get establishment by date
      description: Get establishments created at the indicated date. The date format is timestamp in miliseconds
      operationId: getEstablishmentByDate
      parameters:
        - $ref: '#/components/parameters/date'
      responses:
        200:
          description: Ok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Establishment'
        400:
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        401:
          description: Unauthorize exception
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
  /establishments/date/between:
    get:
      tags:
        - Establishment
      summary: Get establishment by date between
      description: Get establishments created at the indicated dates. The dates format is timestamp in miliseconds
      operationId: getEstablishmentByDateBetween
      parameters:
        - $ref: '#/components/parameters/minDate'
        - $ref: '#/components/parameters/maxDate'
      responses:
        200:
          description: Ok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Establishment'
        400:
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        401:
          description: Unauthorize exception
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
  /establishments/{id}/score:
    patch:
      security:
        - bearerAuth: []
      tags:
        - Establishment
      summary: Update establishment score
      description: Update the score of the establishment indicated.
      operationId: updateEstablishmentScore
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        200:
          description: Ok
          content:
            application/json:
              schema:
                type: string
        401:
          description: Unauthorize exception
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        404:
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
  /establishments/score/{score}:
    get:
      tags:
        - Establishment
      summary: Get establishment by score
      description: Get establishments with the score indicated. The score goes from 0 to 5.
      operationId: getEstablishmentByScore
      parameters:
        - $ref: '#/components/parameters/score'
      responses:
        200:
          description: Ok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Establishment'
        400:
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        401:
          description: Unauthorize exception
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
...
components:
  schemas:
...
    Establishment:
      properties:
        id:
          type: integer
          format: int64
          example: 0
        name:
          type: string
          example: Casa Paco
        creationDate:
          type: string
          format: dd-MM-yyyy
          example: 05-04-2021
        location:
          $ref: '#/components/schemas/Location'
        open:
          type: boolean
          example: false
        score:
          type: number
          example: 4.5            
    EstablishmentDTO:
      properties:
        name:
          type: string
          example: Casa Paco
        location:
          $ref: '#/components/schemas/Location'
        open:
          type: boolean
          example: false          
    Location:
      properties:
        latitude:
          type: number
          example: 0.0
        longitude:
          type: number
          example: 0.0           
...
    Exception:
      properties:
        code: 
          type: string
        errors:
          type: object
          additionalProperties:
            type: string
        message:
          type: string
  parameters:
    id:
      name: id
      in: path
      description: id of the element
      required: true
      schema:
        type: integer
        format: int64
        example: 0
    date:
      name: date
      in: path
      description: creation date
      required: true
      schema:
        type: integer
        example: 1640604419000
    minDate:
      name: minDate
      in: query
      description: date from in timestamp miliseconds
      required: true
      schema:
        type: integer
        example: 1640604419000
    maxDate:
      name: maxDate
      in: query
      description: date to in timestamp miliseconds
      required: true
      schema:
        type: integer
        example: 1641604419000
...
    score:
      name: score
      in: path
      required: true
      schema:
        type: number
        example: 1.5
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT